requirements.txt
Django==3.0.8
django_funky_sheets==0.2.0
djangorestframework==3.11.0
